# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: 7Cav API
  description: >-
    Read-only access to 7Cav.us milpacs (military personnel records) and forum
    tickets. All endpoints require a bearer API key.
  version: dev
...
servers:
  - url: https://api.7cav.us
...
security:
  - bearer: []
paths:
  /api/v1/milpacs/profile/id/{userId}:
    get:
      operationId: MilpacService_GetProfile
      summary: Get given user milpac Profile
      description: >-
        Get milpac Profile data for a specific user. The path value is the
        milpac RELATION key, not the forum user id.
      tags: [milpacs]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
            description: 64-bit id in its decimal-string wire form.
        - name: username
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: The full milpac profile.
          headers:
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale —
                consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        "400":
          $ref: '#/components/responses/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Error'
        "404":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/milpacs/profile/username/{username}:
    get:
      operationId: MilpacService_GetProfile2
      summary: Get given user milpac Profile
      description: Get milpac Profile data for a specific user by username.
      tags: [milpacs]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        "200":
          description: The full milpac profile.
          headers:
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale —
                consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/milpac/discord/{discordId}:
    get:
      operationId: MilpacService_GetUserViaDiscordId
      summary: Get a Milpac profile via Discord ID
      description: Get all milpac Profile data for a given user via their Discord ID.
      tags: [milpacs]
      parameters:
        - name: discordId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The full milpac profile.
          headers:
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale —
                consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/milpac/gamertag/{gamertag}:
    get:
      operationId: MilpacService_GetGamertagProfile
      summary: Get a Milpac profile via Gamertag
      description: Get all milpac Profile data for a given user via their console Gamertag.
      tags: [milpacs]
      parameters:
        - name: gamertag
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The full milpac profile.
          headers:
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale —
                consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/roster/{roster}:
    get:
      operationId: MilpacService_GetRoster
      summary: Get a given roster
      description: Get all milpac Profile data for a given roster.
      tags: [milpacs]
      parameters:
        - $ref: '#/components/parameters/RosterPath'
      responses:
        "200":
          description: Profiles keyed by milpac relation id.
          headers:
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale —
                consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Roster'
        "400":
          $ref: '#/components/responses/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/roster/{roster}/lite:
    get:
      operationId: MilpacService_GetLiteRoster
      summary: Get a given roster with minimal profile data
      description: >-
        Get basic milpac Profile data for a given roster, excluding records
        and awards.
      tags: [milpacs]
      parameters:
        - $ref: '#/components/parameters/RosterPath'
      responses:
        "200":
          description: Lite profiles keyed by milpac relation id.
          headers:
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale —
                consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiteRoster'
        "400":
          $ref: '#/components/responses/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/s1/uniforms/{roster}:
    get:
      operationId: MilpacService_GetS1UniformsRoster
      summary: Get a given roster tailored for S1 Uniforms
      description: Get the specified roster with data tailored for the S1 uniforms tool.
      tags: [milpacs]
      parameters:
        - $ref: '#/components/parameters/RosterPath'
      responses:
        "200":
          description: S1 uniforms profiles keyed by milpac relation id.
          headers:
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale —
                consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/S1UniformsRoster'
        "400":
          $ref: '#/components/responses/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/milpacs/position/search/{positionQuery}:
    get:
      operationId: MilpacService_SearchByPosition
      summary: Search for profiles by position title
      description: >-
        Search for profiles by position title substring (SQL LIKE '%query%'
        against primary and secondary-capable position titles); returns lite
        profiles. The live route is a multi-segment glob
        ({position_query=**}, proto/milpacs.proto): slashes inside the query
        and the bare trailing-slash form also reach the handler — forms an
        OpenAPI path template cannot express; this document records the
        canonical single-segment form.
      tags: [milpacs]
      parameters:
        - name: positionQuery
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: >-
            Matching lite profiles keyed by milpac relation id; {} when
            nothing matches (200, never 404).
          headers:
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale —
                consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiteRoster'
        "400":
          $ref: '#/components/responses/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/milpacs/ranks:
    get:
      operationId: MilpacService_GetAllRanks
      summary: Get all ranks
      description: Get a list of all ranks and their details.
      tags: [milpacs]
      responses:
        "200":
          description: Rank reference list.
          headers:
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale —
                consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RanksResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/milpacs/position/groups:
    get:
      operationId: MilpacService_GetPositionGroups
      summary: Get all position groups with positions
      description: >-
        Get a hierarchical list of all position groups and their associated
        positions.
      tags: [milpacs]
      responses:
        "200":
          description: Position group reference tree.
          headers:
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale —
                consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionGroupsResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/milpacs/awol:
    get:
      operationId: MilpacService_GetAwol
      summary: Get all AWOL troopers
      description: Get a list of all troopers that have not posted in the last 7 days.
      tags: [milpacs]
      responses:
        "200":
          description: AWOL list.
          headers:
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale —
                consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AwolResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/tickets:
    get:
      operationId: TicketsService_ListTickets
      summary: List tickets
      description: >-
        List tickets matching optional filters. Cursor-paginated, sorted by
        last_modified_date DESC with ticket_id DESC as the tiebreaker (the
        cursor is stable under the non-unique sort key). Every query key is
        accepted in both
        lowerCamelCase and snake_case spellings (e.g. perPage / per_page);
        unknown query parameters are ignored. Multi-select filters repeat the
        param (?statusId=1&statusId=5); the comma-separated form is NOT
        supported.
      tags: [tickets]
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: array
            items:
              type: integer
              minimum: 0
        - name: excludeSubcategories
          in: query
          required: false
          schema:
            type: boolean
        - name: ticketState
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: statusId
          in: query
          required: false
          schema:
            type: array
            items:
              type: integer
              minimum: 0
        - name: prefixId
          in: query
          required: false
          schema:
            type: array
            items:
              type: integer
              minimum: 0
        - name: assignedUserId
          in: query
          required: false
          schema:
            type: array
            items:
              type: integer
              minimum: 0
        - name: starterUserId
          in: query
          required: false
          schema:
            type: array
            items:
              type: integer
              minimum: 0
        - name: modifiedSince
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
        - name: includeHidden
          in: query
          required: false
          schema:
            type: boolean
        - name: perPage
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
        - name: afterCursor
          in: query
          required: false
          schema:
            type: string
            description: >-
              Opaque cursor; empty string = first page. Round-trip the
              nextCursor from the previous response.
      responses:
        "200":
          description: One page of tickets.
          headers:
            Cache-Control:
              description: >-
                Freshness signal: never cached by the retired response
                cache; always served live (stale immediately).
              schema:
                type: string
                const: max-age=0
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTicketsResponse'
        "400":
          $ref: '#/components/responses/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/tickets/{ticketId}:
    get:
      operationId: TicketsService_GetTicket
      summary: Get a ticket by numeric ID
      description: >-
        Fetch a single ticket by its integer ticket id. Response includes the
        first 10 messages (oldest first) and the total message count; use
        ListTicketMessages for the full thread.
      tags: [tickets]
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: The ticket with its first messages.
          headers:
            Cache-Control:
              description: >-
                Freshness signal: never cached by the retired response
                cache; always served live (stale immediately).
              schema:
                type: string
                const: max-age=0
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTicketResponse'
        "400":
          $ref: '#/components/responses/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
  /api/v1/tickets/categories:
    get:
      operationId: TicketsService_ListCategories
      summary: List ticket categories
      description: >-
        Returns the full nixfifty category tree. Reference data for the
        categoryId filter on ListTickets.
      tags: [tickets]
      responses:
        "200":
          description: Category reference tree.
          headers:
            Cache-Control:
              description: >-
                Freshness signal: never cached by the retired response
                cache; always served live (stale immediately).
              schema:
                type: string
                const: max-age=0
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCategoriesResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    # The #91 fix, inexpressible in Swagger 2.0: http/bearer makes Swagger UI
    # auto-prepend "Bearer " so users paste only the raw cav7_… key.
    bearer:
      type: http
      scheme: bearer
      description: >-
        7Cav API key (cav7_…), sent as `Authorization: Bearer <key>`.
        Requests without the Bearer scheme receive a plain-text 401 naming
        the expected header form; unknown keys receive a generic plain-text
        401; valid keys lacking the operation's scope receive a JSON 403.
  parameters:
...
    RosterPath:
      name: roster
      in: path
      required: true
      schema:
        # The explicit top-level type matters: it routes the value through
        # full JSON-schema validation, so a literal that is neither an enum
        # name nor a number is a request-contract violation (observed 400).
        type: string
        anyOf:
          - $ref: '#/components/schemas/RosterType'
          - type: string
            pattern: '^[0-9]+$'
            description: Numeric form of the RosterType enum.
  responses:
...
    Error:
      description: >-
        gRPC-status JSON error body. code is the numeric gRPC code (3
        InvalidArgument→400, 5 NotFound→404, 7 PermissionDenied→403,
        13 Internal→500); message strings are frozen verbatim.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
  schemas:
...
    RanksResponse:
      type: object
      properties:
        ranks:
          type: array
          items:
            $ref: '#/components/schemas/RankExpanded'
      required: [ranks]
      additionalProperties: false
...